requirements.txt
scrapy~=2.11.2
spacy~=3.7.6
beautifulsoup4~=4.10.0
gitpython~=3.1.24
pymongo~=4.9.1
datasets>=2.16.1
googletrans>=3.1.0a0
transformers>=4.39.3
nltk~=3.7
torch~=2.1.2
